+++ cj12/file.py
@@ -1,0 +1,27 @@
+from collections.abc import Awaitable, Callable
+from typing import Any
+
+from js import FileReader, alert
+
+from cj12.dom import add_event_listener, elem_by_id
+
+
+class FileInput:
+    def __init__(self, on_data_received: Callable[[bytes], Awaitable[None]]) -> None:
+        self._on_data_received = on_data_received
+
+        self._reader = FileReader.new()
+        add_event_listener(self._reader, "load", self._on_data_load)
+        add_event_listener(self._reader, "error", self._on_error)
+        add_event_listener(elem_by_id("file-input"), "change", self._on_file_change)
+
+    def _on_file_change(self, event: Any) -> None:
+        file = event.target.files.item(0)
+        elem_by_id("dropzone").innerText = f"{file.name} ({file.size / 1024:.2f} KB)"
+        self._reader.readAsBinaryString(file)
+
+    async def _on_data_load(self, _: object) -> None:
+        await self._on_data_received(self._reader.result.encode())
+
+    async def _on_error(self, _: object) -> None:
+        alert("Failed to read file")
